requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.10.0
librosa>=0.8.1
opencv-python-headless>=4.5.3
PyYAML>=5.4.1
pytest>=6.0.0
flake8>=3.9.0
black>=21.0
tqdm>=4.60.0
torchvision
